Profile Custom Metadata Updating
This resolves the profile custom metadata updating issue by comparing the custom metadata to the API vs. comparing it to the local file.

This also adds the ability to set the "Install On" value within Kandji for profiles either by using the runs_on key with a comma-separated list of mac, iphone, ipad, or tv. Or can be set by using runs_on_mac = true, etc. with boolean values.