Add library_item_id to metadata
Added the ability to set a library_item_id for a profile or script. This id is then evaluated when making changes. For example, you could set a library item id and then change the name of a profile - rather than git2kandji creating a new profile since the names no longer match like it was originally checking against, it will update the name using the library item ID.

Added the library_item_id to the downloads.
Add the runs_on value to the profile downloads.

Improvements:
Add library_item_id value to newly created scripts or profiles so it doesn't need to be manually added later.