Remove truncation and ERROR when a name is over 50 characters
I didn't properly handle truncation in the first edition and so scripts with names over 50 characters would sync every run. Looking at this more, I don't think its worth the effort at this time to truncate the name. Instead, we're going to generate an error and proceed on with the script.